refactor(workflows): move StreamDisconnectedError into extra/exceptions.py
