+++ src/mistralai/extra/exceptions.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from dataclasses import dataclass
 from enum import Enum
-from typing import Any, Optional, Union, TYPE_CHECKING
+from typing import Any, Literal, Optional, Union, TYPE_CHECKING
 import typing
 
 from mistralai.client.models import (
@@ -30,6 +30,25 @@ class WorkflowPayloadEncryptionException(MistralClientException):
 
 class WorkflowPayloadCompressionException(MistralClientException):
     """Workflow payload compression exception"""
+
+
+StreamDisconnectReason = Literal["read_error", "stream_error", "internal_error"]
+
+
+class StreamDisconnectedError(MistralClientException):
+    """Raised when a workflow SSE stream is terminated by a server error frame.
+
+    The server ends a stream by emitting an ``event: error`` SSE frame. The SDK
+    surfaces this as a raised exception so consumers can wrap stream iteration in
+    ``try`` / ``except`` instead of inspecting each event for ``event == "error"``.
+
+    Both attributes are populated from the frame's ``data`` JSON payload.
+    """
+
+    def __init__(self, *, reason: StreamDisconnectReason, error: str) -> None:
+        self.reason: StreamDisconnectReason = reason
+        self.error = error
+        super().__init__("Workflow stream disconnected by server")
 
 
 class RunException(MistralClientException):
